% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_size.R
\name{est_size}
\alias{est_size}
\title{Expected Sample Sizes in Small Areas}
\usage{
est_size(data1, data2, domain, n1, n2, sim)
}
\arguments{
<<<<<<< HEAD
\item{data1}{Data frame consisting variables: area_1 area membership variable}

\item{data2}{Data frame consisting variables: area_1 area membership variable}
=======
\item{data1}{Data frame consisting variables: area_1 area #'membership variable}

\item{data2}{Data frame consisting variables: area_1 area #'membership variable}
>>>>>>> 4f1fedfd27406e910ad53445f6206defc49fce9e

\item{domain}{a vector consisting list of area}

\item{n1}{sample size selected at first occasion}

\item{n2}{sample size selected at second occasion}

\item{sim}{Number of bootstrap replication to obtain standard error}
}
\value{
The expected sample sizes in different areas
}
\description{
Expected Sample Sizes in Small Areas
}
\examples{
 est_size(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4)), c(1,2,3), 30, 30,100)
}
