% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimate}
\alias{estimate}
\title{Title}
\usage{
estimate(
  data1,
  data2,
  domain,
  n1,
  n2,
  Si = 0.5,
  lmda_1 = 0.5,
  strategy = "direct_s1",
  sim = "500"
)
}
\arguments{
\item{data1}{Data frame consisting var1 area membership variable, var 2 the study variable....}

\item{data2}{Data frame consisting var1 area membership variable, var 2 the study variable....}

\item{domain}{a vector consisting list of areas in numeric form}

\item{n1}{sample size selected at first occasion}

\item{n2}{sample size selected at second occasion}

\item{Si}{The weight parameter for synthetic estimator  (by defualt si=0.5)}

\item{lmda_1}{The weight parameter used in Strategy 2 and Strategy 3 (by default lmda_1=0.5)}

\item{strategy}{The strategy used for estimation  (by default it will provide result under direct_s1)}

\item{sim}{Number of bootstrap replication to obtain standard error}
}
\value{
The values of estimates, standard errors and 95\% C.I
}
\description{
Title
}
\examples{
 
#Example 1: Result with strategy=direct_s1 
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4)), c(1,2,3),
30, 30,0.3,0.5, strategy='direct_s1',100)
#Example 2 Result with strategy=direct_s1  
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='direct_s2',100)
#Example 3 Result with strategy=weighted_s1  
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='weighted_s1',100)
#Example 4 Result with strategy=weighted_s2 
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='weighted_s2',100)
#Example 5 Result with strategy=poststratified_s1  
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), weight_1=runif(100), stratum_1=c(rep(1,50), rep(2,50))),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), weight_2=runif(100), stratum_2=c(rep(1,50), rep(2,50))), c(1,2,3), 30, 30,0.3,0.5, strategy='poststratified_S1',100)
#Example 6 Result with strategy=poststratified_s2 
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), weight_1=runif(100), stratum_1=c(rep(1,50), rep(2,50))),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), weight_2=runif(100), stratum_2=c(rep(1,50), rep(2,50))), c(1,2,3), 30, 30,0.3,0.5, strategy='poststratified_S2',100)
#Example 7 Result with strategy=regression_s2
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='regression_s2',100)
#Example 8 Result with strategy=ratio_s2
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='ratio_s2',100)
#Example 9 Result with strategy=composite_regression_s2
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='composite_regression_s2',100)
#Example 10 Result with strategy=composite_ratio_s2
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='composite_ratio_s2',100)
#Example 11 Result with strategy=regression_s3
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='regression_s3',100)
#Example 12 Result with strategy=ratio_s3
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='ratio_s3',100)
#Example 13 Result with strategy=composite_regression_s3
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='composite_regression_s3',100)
#Example 14 Result with strategy=composite_ratio_s3
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='composite_ratio_s3',100)
#Example 15 Result with strategy=regression_s4
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='regression_s4',100)
#Example 16 Result with strategy=ratio_s4
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='ratio_s4',100)
#Example 17 Result with strategy=composite_regression_s4
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='composite_regression_s4',100)
#Example 18 Result with strategy=composite_ratio_s4
estimate(data.frame(area_1=c(rep(1,60),rep(2,20),rep(3,20)), Y_1=rnorm(100,4), X_1=rnorm(100,10), weight_1=runif(100)),
data.frame(area_2=c(rep(1,30),rep(2,40),rep(3,30)), Y_2=rnorm(100,4), X_2=rnorm(100,10), weight_2=runif(100)), c(1,2,3), 30, 30,0.3,0.5, strategy='composite_ratio_s4',100)
}
